+++ v2/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Change Log
 
 ## [master](https://github.com/arangodb/go-driver/tree/master) (N/A)
+
+## [2.1.2](https://github.com/arangodb/go-driver/tree/v2.1.2) (2024-11-15)
 - Expose `NewType` method
 - Connection configuration helper
 - Adjust Cursor options
 - Switch to Go 1.22.8
+- Remove deprecated context functions
+- Fix Error Handler in CreateCollectionWithOptions
 
 ## [2.1.1](https://github.com/arangodb/go-driver/tree/v2.1.1) (2024-09-27)
 - Improve backup tests stability